1.管理员登录注册 2.周报添加和查看功能

+++ src/main/java/com/zzsys/threet/repository/StudentRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 public interface StudentRepository extends JpaRepository<Student,Long> {
     Student findByName(String name);
 
-    Student findByNameAndId(String name, int id);
+    Student findByNameAndId(String name, long id);
 
 
 }
